requirements.txt
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
echo "##active_line2##"
numpy
echo "##active_line3##"
torch
echo "##active_line4##"
pydub
